+++ src/CsrfViewInjection.php
@@ -10,10 +10,7 @@ use Yiisoft\Csrf\CsrfTokenInterface;
 /**
  * CsrfViewInjection injects the necessary data into the view to protect against a CSRF attack.
  */
-final class CsrfViewInjection implements
-    ContentParametersInjectionInterface,
-    LayoutParametersInjectionInterface,
-    MetaTagsInjectionInterface
+final class CsrfViewInjection implements CommonParametersInjectionInterface, MetaTagsInjectionInterface
 {
     public const DEFAULT_META_ATTRIBUTE_NAME = 'csrf';
     public const DEFAULT_PARAMETER_NAME = 'csrf';
@@ -60,15 +57,7 @@ final class CsrfViewInjection implements
     /**
      * @throws LogicException when CSRF token is not defined
      */
-    public function getContentParameters(): array
-    {
-        return [$this->parameterName => $this->csrfToken->getValue()];
-    }
-
-    /**
-     * @throws LogicException when CSRF token is not defined
-     */
-    public function getLayoutParameters(): array
+    public function getCommonParameters(): array
     {
         return [$this->parameterName => $this->csrfToken->getValue()];
     }