+++ tests/views/override-layout-parameters/content.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * @var \Yiisoft\View\WebView $this
+ */
+
+$this->setParameter('seoTitle', 'RENDER');